Wildlife Motion Sensor Camera
This is the main software for the my Camera. It is a Python server script that captures a video stream from a Pi Camera and serves it as a .mjpg through a control website to another device. The website can be used to start a photo capture or video capture based on motion detected in the frame. The software is designed to run on a Raspberry Pi Zero W.